x64_requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
filelock==3.19.1
flask-cors==6.0.2
Flask==3.1.3
fsspec==2025.10.0
hf-xet==1.4.2
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==26.0
pillow==11.3.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.7
zipp==3.23.0
torch==2.2.2
